requirements.txt
# === FastAPI & Server ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# === Machine Learning ===
ultralytics>=8.0.0

# === OCR & QR ===
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pyzbar>=0.1.9
openpyxl>=3.1.0

# === Computer Vision ===
opencv-python>=4.8.0
numpy>=1.24.0

# === Face Verification ===
insightface>=0.7.3
onnxruntime>=1.16.0

# === Database (MySQL) ===
PyMySQL>=1.1.0
DBUtils>=3.0.3
python-dotenv>=1.0.0

# === Authentication & Security ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# === Testing ===
pytest>=7.0.0
httpx>=0.25.0
